fix(ebpf): 修复 eBPF verifier 兼容性（5 类问题，6.8 kernel 实测通过）
- 组合栈超限（544>512）：主帧瘦身（删除 dst_key、parse 直接写 pc 字段省 ~72B）、syn cookie 链帧分离与内联优化
- pointer arithmetic with <<=：深层函数不再接收 ctx 入参（LLVM 参数提升触发入参指针移位被 verifier 拒绝），ctx 改从 pc 内存读取
- 标量内存访问：pc.data 丢失 PTR_TO_PACKET 类型，回退 pc.ctx + ptr_at 模式
- for-range 循环 u32 零扩展指令：全部改 u64 while 循环（port_acl/l7_scan/log_packet_sample）
- MSS option 写入越界：边界证明改用 ptr_at_mut 返回值传播（防 LLVM 删除存在性检查）
- syn cookie 恢复的挑战路径在 Ubuntu 6.8 实测：verifier 通过、XDP 挂载运行、防护项目 DROP 生效、正常连接直通

+++ eshield-ebpf/src/l7_scan.rs
@@ -45,9 +45,9 @@ pub fn scan(ctx: &XdpContext, src: &IpKey, ip_hdr_len: usize, protocol: u8, dpor
     // SAFETY：上面的边界检查保证了 8 字节可读。
     let chunk = unsafe { *payload };
 
-    let mut i: u32 = 0;
-    while i < MAX_PATTERNS {
-        let pat = match L7_PATTERNS.get(i) {
+    let mut i: u64 = 0;
+    while i < MAX_PATTERNS as u64 {
+        let pat = match L7_PATTERNS.get(i as u32) {
             Some(p) => p,
             None => {
                 i += 1;
